Move template logic into a separate file
